refactor: 1단계 코드리뷰 반영
