+++ src/main/kotlin/string/add/calculator/StringAddCalculator.kt
@@ -1,6 +1,10 @@
 package string.add.calculator
 
-class StringAddCalculator {
+object StringAddCalculator {
+    private const val COMMA = ","
+    private const val COLON = ":"
+    private val CUSTOM_DELIMITER_PATTERN = "//(.)\n(.*)".toRegex()
+
     fun calculate(text: String?): Int {
         if (text.isNullOrBlank()) {
             return 0
@@ -11,39 +15,11 @@ class StringAddCalculator {
     }
 
     private fun split(text: String): List<String> {
-        val regex = CUSTOM_DELIMITER.toRegex()
-        val matchResult = regex.matchEntire(text)
+        val matchResult = CUSTOM_DELIMITER_PATTERN.matchEntire(text)
         if (matchResult != null) {
             val (delimiter, numbers) = matchResult.destructured
             return numbers.split(delimiter)
         }
         return text.split(COMMA, COLON)
-    }
-
-    companion object {
-        private const val COMMA = ","
-        private const val COLON = ":"
-        private const val CUSTOM_DELIMITER = "//(.)\n(.*)"
-    }
-}
-
-@JvmInline
-value class NaturalNumber(
-    val value: Int,
-) {
-    init {
-        if (value < 0) {
-            throw RuntimeException("음수는 입력할 수 없습니다.")
-        }
-    }
-
-    constructor(value: String) : this(value.toIntOrNull() ?: throw RuntimeException("숫자 이외의 값은 입력할 수 없습니다."))
-
-    fun add(other: NaturalNumber): NaturalNumber = NaturalNumber(value + other.value)
-
-    companion object {
-        val ZERO = NaturalNumber(0)
-
-        fun sum(numbers: List<NaturalNumber>): NaturalNumber = numbers.reduceOrNull { acc, naturalNumber -> acc.add(naturalNumber) } ?: ZERO
     }
 }
