+++ src/test/kotlin/string/add/calculator/StringAddCalculatorTest.kt
@@ -1,7 +1,6 @@
 package string.add.calculator
 
 import io.kotest.matchers.shouldBe
-import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
 import org.junit.jupiter.params.ParameterizedTest
@@ -10,12 +9,7 @@ import org.junit.jupiter.params.provider.ValueSource
 
 @Suppress("NonAsciiCharacters")
 class StringAddCalculatorTest {
-    private lateinit var sut: StringAddCalculator
-
-    @BeforeEach
-    fun setUp() {
-        sut = StringAddCalculator()
-    }
+    private val sut = StringAddCalculator
 
     @ParameterizedTest
     @NullAndEmptySource
